# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sales Program Submission API
  version: 0.3.0
  description: Implemented DRF/MySQL API for pre-provisioned employee login, drafts, sequential approvals, private
    uploads, signed PDFs and review history. Cancellation success remains gated pending allowed source states. Flutter
    business screens and future superadmin/dashboard/recovery APIs are outside this backend phase. See API_DESIGN.md
    and BACKEND.md.
servers:
- url: /api/v1
  description: Relative API root; configure the deployment origin in the client.
security:
- bearerAuth: []
...
paths:
...
  /program-submissions:
    get:
      operationId: listMySubmissions
      tags:
      - Program Submission
      summary: List and search only submissions owned by the caller.
      description: Submitter role plus owner scope. Apply filters before pagination. Search program name/number; execution
        date filters bound periodStart inclusively. OR within repeated filters, AND between filters. Stable sorting
        appends id ascending. Out-of-range pages are empty.
      x-allowed-roles: submitter (owner)
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          maximum: 2147483647
        description: 1-based page.
        example: 1
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
        description: Items per page.
        example: 20
      - name: q
        in: query
        required: false
        schema:
          type: string
          maxLength: 100
          minLength: 1
        description: Trimmed, case-insensitive literal substring of program number or program name; no wildcard syntax.
        example: Bundling
      - name: programNumber
        in: query
        required: false
        schema:
          type: string
          maxLength: 200
        description: Exact program number; combines with q using AND.
        example: PRG-2026-0144
      - name: status
        in: query
        required: false
        schema:
          type: array
          items:
            $ref: '#/components/schemas/SubmissionStatus'
          maxItems: 100
        description: Repeat key for multiple statuses; OR within this filter.
        style: form
        explode: true
      - name: locationId
        in: query
        required: false
        schema:
          type: array
          items:
            $ref: '#/components/schemas/Id'
          maxItems: 100
        description: Repeat key; submission must match at least one selected location.
        style: form
        explode: true
      - name: programTypeId
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/Id'
        description: Exact type ID; submissions with unknown type do not match.
      - name: periodStartFrom
        in: query
        required: false
        schema:
          type: string
          format: date
        description: Inclusive lower bound on execution periodStart; not submittedAt.
      - name: periodStartTo
        in: query
        required: false
        schema:
          type: string
          format: date
        description: Inclusive upper bound on execution periodStart; lower bound must not exceed upper bound.
      - name: sort
        in: query
        required: false
        schema:
          type: string
          enum:
          - -createdAt
          - createdAt
          - -updatedAt
          - updatedAt
          - programNumber
          default: -createdAt
        description: Default -createdAt; append id ascending internally as a stable tie-breaker.
        example: -createdAt
      responses:
        '200':
          description: Successful operation.
          headers:
            X-Request-Id:
              schema:
                type: string
              description: Server-generated correlation ID.
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionListResponse'
              example:
                data:
                - id: sub_0144
                  programNumber: PRG-2026-0144
                  programName: Bundling Idul Adha Outlet BSD
                  programType:
                    id: typ_bundling
                    code: bundling
                    name: Bundling
                    isActive: true
                  locations:
                  - id: loc_bsd
                    code: BSD
                    name: BSD
                    isActive: true
                  periodStart: '2026-09-06'
                  periodEnd: '2026-09-20'
                  estimatedCost:
                    currency: IDR
                    amount: '42000000.00'
                  owner:
                    id: usr_rizky
                    fullName: Rizky Pratama
                    jobTitle: Sales Executive
                  status: draft
                  currentStage: null
                  createdAt: '2026-08-21T03:00:00Z'
                  updatedAt: '2026-08-21T03:00:00Z'
                  submittedAt: null
                  version: 1
                  myActiveTaskIds: []
                meta:
                  requestId: req_01salesdesign
                  page: 1
                  pageSize: 20
                  totalItems: 1
                  totalPages: 1
                  hasNextPage: false
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '422':
          $ref: '#/components/responses/422'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'
        '405':
          $ref: '#/components/responses/405'
        '406':
          $ref: '#/components/responses/406'
    post:
      operationId: createDraft
      tags:
      - Program Submission
      summary: Create a persisted draft, including an empty draft.
      description: Submitter role required. {} creates an empty draft. All supplied references must be active and
        authorized; selected reviewers require explicit per-employee eligibility plus the matching role. Allocate
        PRG-YYYY-NNNN under a MySQL row lock, yearly in Asia/Jakarta. Owner, checker, number, status, timestamps and
        tasks are server-owned. Never derive type/cost from uploaded spreadsheets. Drafts may be incomplete. Inspect
        submissionIssues before submit.
      x-allowed-roles: submitter
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 16
          maxLength: 128
        description: Random key for one logical operation; reuse only for an identical retry. Retained for 24 hours.
        example: sales-example-key-0001
      responses:
        '201':
          description: Successful operation.
          headers:
            X-Request-Id:
              schema:
                type: string
              description: Server-generated correlation ID.
            Cache-Control:
              schema:
                type: string
                const: no-store
            ETag:
              schema:
                type: string
                pattern: ^"[1-9][0-9]*"$
              description: Quoted parent submission version.
            Location:
              schema:
                type: string
              description: Relative URL of the created resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionResponse'
              example:
                data:
                  id: sub_0144
                  programNumber: PRG-2026-0144
                  programName: Bundling Idul Adha Outlet BSD
                  programType:
                    id: typ_bundling
                    code: bundling
                    name: Bundling
                    isActive: true
                  locations:
                  - id: loc_bsd
                    code: BSD
                    name: BSD
                    isActive: true
                  periodStart: '2026-09-06'
                  periodEnd: '2026-09-20'
                  estimatedCost:
                    currency: IDR
                    amount: '42000000.00'
                  owner:
                    id: usr_rizky
                    fullName: Rizky Pratama
                    jobTitle: Sales Executive
                  status: draft
                  currentStage: null
                  createdAt: '2026-08-21T03:00:00Z'
                  updatedAt: '2026-08-21T03:00:00Z'
                  submittedAt: null
                  version: 1
                  myActiveTaskIds: []
                  reviewPlan:
                    checker:
                      id: usr_andi
                      fullName: Andi Setiawan
                      jobTitle: Supervisor Sales
                    acknowledgers:
                    - id: usr_dewi
                      fullName: Dewi Larasati
                      jobTitle: Branch Manager BSD
                    approvers:
                    - id: usr_ratna
                      fullName: Ratna Kusuma
                      jobTitle: GM Sales & Marketing
                  reviewTasks: []
                  attachments: []
                  allowedActions:
                  - update
                  - uploadAttachment
                  - downloadPdf
                  submissionIssues:
                  - field: null
                    code: WORKFLOW_POLICY_UNRESOLVED
                    message: Review policy must be confirmed before submission.
                meta:
                  requestId: req_01salesdesign
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '409':
          $ref: '#/components/responses/409'
        '415':
          $ref: '#/components/responses/415'
        '422':
          $ref: '#/components/responses/422'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'
        '405':
          $ref: '#/components/responses/405'
        '406':
          $ref: '#/components/responses/406'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftCreate'
            example:
              programName: Bundling Idul Adha Outlet BSD
              locationIds:
              - loc_bsd
              periodStart: '2026-09-06'
              periodEnd: '2026-09-20'
              acknowledgerIds:
              - usr_dewi
              approverIds:
              - usr_ratna
              programTypeId: typ_bundling
              estimatedCost:
                currency: IDR
                amount: '42000000.00'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Opaque short-lived access token. Roles and object scope are enforced by server policy, not this
        bearer declaration.
  schemas:
    Id:
      type: string
      pattern: ^[a-z][a-z0-9_]{2,63}$
      description: Opaque, immutable ID; clients must not parse its contents.
...
    Stage:
      type: string
      enum:
      - checker
      - acknowledgement
      - approval
    SubmissionStatus:
      type: string
      enum:
      - draft
      - pendingChecker
      - pendingAcknowledgement
      - pendingApproval
      - approved
      - rejected
      - cancelled
    Meta:
      type: object
      additionalProperties: false
      properties:
        requestId:
          type: string
          minLength: 1
      required:
      - requestId
    PageMeta:
      type: object
      additionalProperties: false
      properties:
        requestId:
          type: string
          minLength: 1
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
        totalItems:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        hasNextPage:
          type: boolean
      required:
      - requestId
      - page
      - pageSize
      - totalItems
      - totalPages
      - hasNextPage
    Error:
      type: object
      additionalProperties: false
      properties:
        error:
          type: object
          additionalProperties: false
          properties:
            code:
              type: string
              minLength: 1
            message:
              type: string
              minLength: 1
            details:
              type: array
              items:
                type: object
                additionalProperties: false
                properties:
                  field:
                    anyOf:
                    - type: string
                    - type: 'null'
                  code:
                    type: string
                  message:
                    type: string
                required:
                - field
                - code
                - message
          required:
          - code
          - message
          - details
        meta:
          $ref: '#/components/schemas/Meta'
      required:
      - error
      - meta
    Person:
      type: object
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/Id'
        fullName:
          type: string
          minLength: 1
          maxLength: 150
        jobTitle:
          anyOf:
          - type: string
            maxLength: 150
          - type: 'null'
      required:
      - id
      - fullName
      - jobTitle
...
    Location:
      type: object
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/Id'
        code:
          type: string
        name:
          type: string
        isActive:
          type: boolean
      required:
      - id
      - code
      - name
      - isActive
    ProgramType:
      type: object
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/Id'
        code:
          type: string
        name:
          type: string
        isActive:
          type: boolean
      required:
      - id
      - code
      - name
      - isActive
    Money:
      type: object
      additionalProperties: false
      properties:
        currency:
          type: string
          enum:
          - IDR
        amount:
          type: string
          pattern: ^(0|[1-9][0-9]{0,13})(\.[0-9]{2})$
          description: Decimal major units, never a floating point number; display formatting is client-side.
      required:
      - currency
      - amount
...
    SubmissionSummary:
      type: object
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/Id'
        programNumber:
          type: string
        programName:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
        programType:
          anyOf:
          - $ref: '#/components/schemas/ProgramType'
          - type: 'null'
        locations:
          type: array
          items:
            $ref: '#/components/schemas/Location'
        periodStart:
          anyOf:
          - type: string
            format: date
          - type: 'null'
        periodEnd:
          anyOf:
          - type: string
            format: date
          - type: 'null'
        estimatedCost:
          anyOf:
          - $ref: '#/components/schemas/Money'
          - type: 'null'
        owner:
          $ref: '#/components/schemas/Person'
        status:
          $ref: '#/components/schemas/SubmissionStatus'
        currentStage:
          anyOf:
          - $ref: '#/components/schemas/Stage'
          - type: 'null'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        submittedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
        version:
          type: integer
          minimum: 1
        myActiveTaskIds:
          type: array
          items:
            $ref: '#/components/schemas/Id'
      required:
      - id
      - programNumber
      - programName
      - programType
      - locations
      - periodStart
      - periodEnd
      - estimatedCost
      - owner
      - status
      - currentStage
      - createdAt
      - updatedAt
      - submittedAt
      - version
      - myActiveTaskIds
...
    DraftCreate:
      type: object
      additionalProperties: false
      properties:
        programName:
          anyOf:
          - type: string
            minLength: 1
            maxLength: 200
          - type: 'null'
        locationIds:
          type: array
          items:
            $ref: '#/components/schemas/Id'
          uniqueItems: true
          maxItems: 100
        periodStart:
          anyOf:
          - type: string
            format: date
          - type: 'null'
        periodEnd:
          anyOf:
          - type: string
            format: date
          - type: 'null'
        acknowledgerIds:
          type: array
          items:
            $ref: '#/components/schemas/Id'
          uniqueItems: true
          maxItems: 4
        approverIds:
          type: array
          items:
            $ref: '#/components/schemas/Id'
          uniqueItems: true
          maxItems: 5
        programTypeId:
          anyOf:
          - $ref: '#/components/schemas/Id'
          - type: 'null'
        estimatedCost:
          anyOf:
          - $ref: '#/components/schemas/Money'
          - type: 'null'
      required: []
...
    SubmissionResponse:
      type: object
      additionalProperties: false
      properties:
        data:
          $ref: '#/components/schemas/Submission'
        meta:
          $ref: '#/components/schemas/Meta'
      required:
      - data
      - meta
    SubmissionListResponse:
      type: object
      additionalProperties: false
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/SubmissionSummary'
        meta:
          $ref: '#/components/schemas/PageMeta'
      required:
      - data
      - meta
...
  responses:
    '400':
      description: Malformed JSON, unknown parameters, invalid query encoding or missing Idempotency-Key.
      headers:
        X-Request-Id:
          schema:
            type: string
          description: Correlation ID also returned in meta.requestId.
        Cache-Control:
          schema:
            type: string
            const: no-store
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: BAD_REQUEST
              message: Malformed JSON, unknown parameters, invalid query encoding or missing Idempotency-Key.
              details: []
            meta:
              requestId: req_01salesdesign
    '401':
      description: Credentials or session are invalid or expired.
      headers:
        X-Request-Id:
          schema:
            type: string
          description: Correlation ID also returned in meta.requestId.
        Cache-Control:
          schema:
            type: string
            const: no-store
        WWW-Authenticate:
          schema:
            type: string
          description: Bearer challenge; never includes credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHENTICATED
              message: Credentials or session are invalid or expired.
              details: []
            meta:
              requestId: req_01salesdesign
    '403':
      description: The authenticated user lacks the required role.
      headers:
        X-Request-Id:
          schema:
            type: string
          description: Correlation ID also returned in meta.requestId.
        Cache-Control:
          schema:
            type: string
            const: no-store
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: FORBIDDEN
              message: The authenticated user lacks the required role.
              details: []
            meta:
              requestId: req_01salesdesign
...
    '409':
      description: The operation conflicts with the current resource state.
      headers:
        X-Request-Id:
          schema:
            type: string
          description: Correlation ID also returned in meta.requestId.
        Cache-Control:
          schema:
            type: string
            const: no-store
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: CONFLICT
              message: The operation conflicts with the current resource state.
              details: []
            meta:
              requestId: req_01salesdesign
...
    '415':
      description: The request or file media type is unsupported.
      headers:
        X-Request-Id:
          schema:
            type: string
          description: Correlation ID also returned in meta.requestId.
        Cache-Control:
          schema:
            type: string
            const: no-store
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNSUPPORTED_MEDIA_TYPE
              message: The request or file media type is unsupported.
              details: []
            meta:
              requestId: req_01salesdesign
    '422':
      description: One or more fields violate validation rules.
      headers:
        X-Request-Id:
          schema:
            type: string
          description: Correlation ID also returned in meta.requestId.
        Cache-Control:
          schema:
            type: string
            const: no-store
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: VALIDATION_FAILED
              message: One or more fields violate validation rules.
              details: []
            meta:
              requestId: req_01salesdesign
...
    '429':
      description: Too many requests. Retry after the indicated delay.
      headers:
        X-Request-Id:
          schema:
            type: string
          description: Correlation ID also returned in meta.requestId.
        Cache-Control:
          schema:
            type: string
            const: no-store
        Retry-After:
          schema:
            type: integer
            minimum: 1
          description: Delay in seconds.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: RATE_LIMITED
              message: Too many requests. Retry after the indicated delay.
              details: []
            meta:
              requestId: req_01salesdesign
    '500':
      description: An unexpected error occurred.
      headers:
        X-Request-Id:
          schema:
            type: string
          description: Correlation ID also returned in meta.requestId.
        Cache-Control:
          schema:
            type: string
            const: no-store
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: INTERNAL_ERROR
              message: An unexpected error occurred.
              details: []
            meta:
              requestId: req_01salesdesign
    '503':
      description: 'SERVICE_UNAVAILABLE: database, scanner-dependent operation or private file storage is unavailable;
        Retry-After when supplied.'
      headers:
        X-Request-Id:
          schema:
            type: string
          description: Correlation ID also returned in meta.requestId.
        Cache-Control:
          schema:
            type: string
            const: no-store
        Retry-After:
          schema:
            type: integer
            minimum: 1
          description: Delay in seconds.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: SERVICE_UNAVAILABLE
              message: A required service is temporarily unavailable.
              details: []
            meta:
              requestId: req_01salesdesign
    '405':
      description: METHOD_NOT_ALLOWED
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    '406':
      description: NOT_ACCEPTABLE
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...